清理 OptiScaler 构建警告

+++ src/OptiScaler/low_latency/input/input_uell.cpp
@@ -41,7 +41,7 @@ void InputUeLowLatency::tickStart(int64_t frameId, float DeltaSeconds, bool bIdl
     if (!inited)
         InputUeLowLatency::init();
 
-    auto result = InputCommon::sleep(inputContext, device, frameId);
+    auto result = InputCommon::sleep(inputContext, device, static_cast<uint32_t>(frameId));
 
     if (result == InputResult::UsingDifferentInput)
         return;
